+++ src/main/java/com/github/theredbrain/spellengineextension/SpellEngineExtension.java
@@ -8,11 +8,14 @@ import com.github.theredbrain.spellengineextension.compat.RangedWeaponAPICompati
 import com.github.theredbrain.spellengineextension.compat.StaminaAttributesCompatibility;
 import com.github.theredbrain.spellengineextension.component.type.HasConditionalSpellContainerComponent;
 import com.github.theredbrain.spellengineextension.config.ServerConfig;
+import com.github.theredbrain.spellengineextension.enchantment.ProvideSpellsEnchantmentEntityEffect;
 import com.github.theredbrain.spellengineextension.predicate.item.SpellContainersPredicate;
 import com.github.theredbrain.spellengineextension.registry.DataComponentPredicateRegistry;
+import com.github.theredbrain.spellengineextension.registry.EnchantmentEntityEffectRegistry;
 import com.github.theredbrain.spellengineextension.registry.ItemComponentRegistry;
 import com.github.theredbrain.spellengineextension.registry.ServerEventRegistry;
 import com.github.theredbrain.spellengineextension.registry.SpellSchoolRegistry;
+import com.mojang.serialization.MapCodec;
 import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
 import me.fzzyhmstrs.fzzy_config.api.RegisterType;
 import net.fabricmc.api.ModInitializer;
@@ -65,6 +68,8 @@ public class SpellEngineExtension implements ModInitializer {
 	public static ComponentType<HasConditionalSpellContainerComponent> HAS_CONDITIONAL_SPELL_CONTAINER;
 
 	public static ItemSubPredicate.Type<SpellContainersPredicate> SPELL_CONTAINER_PREDICATE;
+
+	public static MapCodec<ProvideSpellsEnchantmentEntityEffect> PROVIDE_SPELLS;
 
 	public static final TagKey<Spell> CAN_BE_IN_USE_ITEM_SPELL_HOTBAR_SLOT = TagKey.of(SpellRegistry.KEY, SpellEngineExtension.identifier("can_be_in_use_item_spell_hotbar_slot"));
 	public static final TagKey<Spell> ENABLES_MOVEMENT_LOCKING_DURING_CASTING = TagKey.of(SpellRegistry.KEY, SpellEngineExtension.identifier("enables_movement_locking_during_casting"));
@@ -158,6 +163,7 @@ public class SpellEngineExtension implements ModInitializer {
 
 		DataComponentPredicateRegistry.init();
 		ItemComponentRegistry.init();
+		EnchantmentEntityEffectRegistry.init();
 		ServerEventRegistry.init();
 		SpellSchoolRegistry.init();