+++ src/main/java/com/github/theredbrain/spellengineextension/mixin/server/network/ServerPlayerEntityMixin.java
@@ -1,6 +1,8 @@
 package com.github.theredbrain.spellengineextension.mixin.server.network;
 
+import com.github.theredbrain.spellengineextension.SpellEngineExtension;
 import com.github.theredbrain.spellengineextension.entity.effect.ProvidesSpell;
+import com.github.theredbrain.spellengineextension.server.network.DuckServerPlayerEntityMixin;
 import com.mojang.authlib.GameProfile;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.effect.StatusEffectInstance;
@@ -8,6 +10,7 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.spell_engine.api.spell.container.SpellContainer;
 import net.spell_engine.internals.container.SpellContainerSource;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -15,11 +18,33 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Mixin(ServerPlayerEntity.class)
-public abstract class ServerPlayerEntityMixin extends PlayerEntity {
+public abstract class ServerPlayerEntityMixin extends PlayerEntity implements DuckServerPlayerEntityMixin {
+
+	@Unique
+	private final List<String> enchantmentProvidedSpellsList = new ArrayList<>();
+
+	@Unique
+	private int oldEnchantmentProvidedSpellsListHash = 0;
 
 	public ServerPlayerEntityMixin(World world, BlockPos pos, float yaw, GameProfile gameProfile) {
 		super(world, pos, yaw, gameProfile);
+	}
+
+	@Inject(method = "tick", at = @At("TAIL"))
+	public void spellengineextension$tick(CallbackInfo ci) {
+		if (this.enchantmentProvidedSpellsList.hashCode() != this.oldEnchantmentProvidedSpellsListHash) {
+			List<String> newSpellList = new ArrayList<>(this.enchantmentProvidedSpellsList);
+
+			((SpellContainerSource.Owner) this).serverSideSpellContainers().put("enchantment_provided", new SpellContainer(SpellContainer.ContentType.NONE, "", "", newSpellList.size(), newSpellList));
+			((SpellContainerSource.Owner) this).markServerSideSpellContainersDirty();
+
+			this.oldEnchantmentProvidedSpellsListHash = this.enchantmentProvidedSpellsList.hashCode();
+		}
+		this.enchantmentProvidedSpellsList.clear();
 	}
 
 	@Inject(method = "onStatusEffectApplied", at = @At("TAIL"))
@@ -35,6 +60,15 @@ public abstract class ServerPlayerEntityMixin extends PlayerEntity {
 	@Inject(method = "onStatusEffectRemoved", at = @At("TAIL"))
 	protected void spellengineextension$onStatusEffectRemoved(StatusEffectInstance effect, CallbackInfo ci) {
 		spellengineextension$clearSpellContainerCache(((ServerPlayerEntity) (Object) this));
+	}
+
+	@Override
+	public void spellengineextension$addEnchantmentProvidedSpells(List<String> newSpellIds) {
+		for (String newSpellId : newSpellIds) {
+			if (!this.enchantmentProvidedSpellsList.contains(newSpellId)) {
+				this.enchantmentProvidedSpellsList.add(newSpellId);
+			}
+		}
 	}
 
 	@Unique