+++ src/main/java/com/github/theredbrain/spellengineextension/server/network/DuckServerPlayerEntityMixin.java
@@ -1,0 +1,8 @@
+package com.github.theredbrain.spellengineextension.server.network;
+
+import java.util.List;
+
+public interface DuckServerPlayerEntityMixin {
+
+	void spellengineextension$addEnchantmentProvidedSpells(List<String> newSpellIds);
+}